feat: 缓存已下载好的Chromium，防止重复下载，然后导致下载错误

+++ index.js
@@ -1,3 +1,4 @@
+const fs = require('fs-extra');
 const path = require('path');
 const chalk = require('chalk');
 const Chromium = require('./chromium');
@@ -25,22 +26,43 @@ process.on('unhandledRejection', (reason, p) => {
 
 // 运行时检查是已安装Chromium
 if (Chromium.isExist === false) {
-  console.error(
-    `Please make sure ${chalk.green('chromium')} have install at ${chalk.yellow(
-      path.join(config.paths.puppeteer, '.local-chromium')
-    )}`
-  );
-  console.error(
-    `Try to reinstall: ${chalk.green('node ' + path.join(config.paths.puppeteer, 'install.js'))}\n`
-  );
+  // 查看本地是否有缓存
+  if (Chromium.isExistLocalCache === true) {
+    fs.copySync(Chromium.cacheChromiumPath, Chromium.localChromiumPath);
+  } else {
+    // 如果没有缓存，则直接退出
+    console.error(
+      `Please make sure ${chalk.green('chromium')} have install at ${chalk.yellow(
+        path.join(config.paths.puppeteer, '.local-chromium')
+      )}`
+    );
+    console.error(
+      `Try to reinstall: ${chalk.green(
+        'node ' + path.join(config.paths.puppeteer, 'install.js')
+      )}\n`
+    );
 
-  console.info(
-    `If you got network trouble, You can download ${chalk.green(
-      Chromium.downloadUrl
-    )} by your self and extract to ${chalk.yellow(Chromium.path)}`
-  );
+    console.info(
+      `If you got network trouble, You can download ${chalk.green(
+        Chromium.downloadUrl
+      )} by your self and extract to ${chalk.yellow(Chromium.path)}`
+    );
 
-  process.exit(1);
+    process.exit(1);
+  }
+} else {
+  // 缓存Chromium，防止重复下载
+  fs.ensureDirSync(Chromium.cacheChromiumPath);
+
+  // 子目录
+  const subPath = Chromium.platform + '-' + Chromium.revision;
+  // 最后缓存的目标路径
+  const targetPath = path.join(Chromium.cacheChromiumPath, subPath);
+
+  // 如果不存在，则缓存起来
+  if (fs.pathExistsSync(targetPath) === false) {
+    fs.copySync(path.join(Chromium.localChromiumPath, subPath), targetPath);
+  }
 }
 
 const App = require('./app/app');
